test(attachments): per-handler sibling tests for boundary convention
